+++ js/main.js
@@ -1,5 +1,5 @@
 
-var version="1.11.1"; 
+var version="1.12.0"; 
 
 var md = new MobileDetect(window.navigator.userAgent);